# swagger/swagger.yaml
swagger: '2.0'
info:
  description: Aplicación dirigida a compañías de Marketing para administrar todos los contactos de sus clientes para sus campañas.
  version: 1.0.0
  title: Data Warehouse
  contact:
    email: solmartinezpisani@gmail.com
...
schemes:
  - https
  - http
paths:
...
  /contactos/{by}:
    get:
      tags:
        - contactos
      summary: Ver contactos
      operationId: getContactos
      produces:
        - application/json
      parameters:
        - name: User token
          in: header
          required: true
          type: string
          description: User JWT token (Bearer)
        - in: path
          name: by
          type: integer
          required: true
          description: The parameter you want your contacts to be ordered by (nombre, interes, cargo, compania, pais)
      responses:
        '200':
          description: Successful operation
          schema:
            type: array
            items:
              $ref: '#/definitions/GetContactos'
        '400':
          description: Unsuccessful
          schema:
            $ref: '#/definitions/Authenticate'
        '404':
          description: Not Found
          schema:
            type: object
            properties:
              status_code:
                type: integer
                format: int64
                example: 404
    patch:
      tags:
        - contactos
      summary: Editar contacto
      operationId: updateContacto
      produces:
        - application/json
      parameters:
        - name: User token
          in: header
          required: true
          type: string
          description: User JWT token (Bearer)
        - in: path
          name: by
          type: integer
          required: true
          description: Numeric ID of the contact to update.
        - in: body
          name: body
          description: >-
            Editar contacto
          required: true
          schema:
            $ref: '#/definitions/CreateContacto'
      responses:
        '200':
          description: Successful operation
          schema:
            properties:
              message:
                type: string
                example: Contacto actualizado
        '400':
          description: Unsuccessful
          schema:
            $ref: '#/definitions/Authenticate'
        '404':
          description: Not Found
          schema:
            type: object
            properties:
              status_code:
                type: integer
                format: int64
                example: 404
    delete: 
      tags:
        - contactos
      summary: Eliminar contacto
      operationId: deleteContacto
      produces:
        - application/json
      parameters:
        - name: User token
          in: header
          required: true
          type: string
          description: User JWT token (Bearer)
        - in: path
          name: by
          type: integer
          required: true
          description: Numeric ID of the contact to delete.
      responses:
        '200':
          description: Successful operation
          schema:
            type: object
            properties:
              message:
                type: string
                example: Contacto eliminado
        '400':
          description: Unsuccessful
          schema:
            $ref: '#/definitions/Authenticate'
        '404':
          description: Not Found
          schema:
            type: object
            properties:
              status_code:
                type: integer
                format: int64
                example: 404 
  /contactos/contacto/{id}:
    get:
      tags:
        - contactos
      summary: Ver un sólo contacto
      operationId: getContacto
      produces:
        - application/json
      parameters:
        - name: User token
          in: header
          required: true
          type: string
          description: User JWT token (Bearer)
        - in: path
          name: id
          type: integer
          required: true
          description: Numeric ID of the contact.
      responses:
        '200':
          description: Successful operation
          schema:
            type: array
            items:
              $ref: '#/definitions/GetContactos'
        '400':
          description: Unsuccessful
          schema:
            $ref: '#/definitions/Authenticate'
        '404':
          description: Not Found
          schema:
            type: object
            properties:
              status_code:
                type: integer
                format: int64
                example: 404
  /contactos: 
    post: 
      tags:
        - contactos
      summary: Crear contacto
      operationId: createContacto
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: Crear contacto
          required: true
          schema:
            $ref: '#/definitions/CreateContacto'
      responses:
        '200':
          description: Successful Register
          schema:
            type: object
            properties:
              message:
                type: string
                example: Contacto creado
        '400':
          description: Unsuccessful Register
          schema:
            type: object
            properties:
              status_code:
                type: integer
                format: int64
                example: 400
              message:
                type: string
                example: Contacto ya guardado
  /contactos/search/{by}:
    post:
      tags:
        - contactos
      summary: Ver contactos de una determinada búsqueda
      operationId: getContactosSearch
      produces:
        - application/json
      parameters:
        - name: User token
          in: header
          required: true
          type: string
          description: User JWT token (Bearer)
        - in: path
          name: by
          type: integer
          required: true
          description: The parameter you want your contacts to be ordered by (nombre, interes, cargo, compania, pais)
        - in: body
          name: body
          description: The information used to search a specific contact
          required: true
          schema:
            $ref: '#/definitions/CreateContacto'
      responses:
        '200':
          description: Successful operation
          schema:
            type: array
            items:
              $ref: '#/definitions/GetContactos'
        '400':
          description: Unsuccessful
          schema:
            $ref: '#/definitions/Authenticate'
        '404':
          description: Not Found
          schema:
            type: object
            properties:
              status_code:
                type: integer
                format: int64
                example: 404
  /contactos/contacto_canales/{id}:
    get:
      tags:
        - contactos
      summary: Ver canales de un contacto
      operationId: getContactoChannels
      produces:
        - application/json
      parameters:
        - name: User token
          in: header
          required: true
          type: string
          description: User JWT token (Bearer)
        - in: path
          name: id
          type: integer
          required: true
          description: Numeric ID of the contact.
      responses:
        '200':
          description: Successful operation
          schema:
            type: array
            items:
              $ref: '#/definitions/GetContactoChannels'
        '400':
          description: Unsuccessful
          schema:
            $ref: '#/definitions/Authenticate'
        '404':
          description: Not Found
          schema:
            type: object
            properties:
              status_code:
                type: integer
                format: int64
                example: 404
  /canales:
    get:
      tags:
        - canales
      summary: Ver todos los canales
      operationId: getChannels
      produces:
        - application/json
      parameters:
        - name: User token
          in: header
          required: true
          type: string
          description: User JWT token (Bearer)
      responses:
        '200':
          description: Successful operation
          schema:
            type: array
            items:
              $ref: '#/definitions/GetChannels'
        '400':
          description: Unsuccessful
          schema:
            $ref: '#/definitions/Authenticate'
        '404':
          description: Not Found
          schema:
            type: object
            properties:
              status_code:
                type: integer
                format: int64
                example: 404
definitions:
  Authenticate:
    type: object
    properties:
      message:
        type: string
        example: Header must contain JWT to verify identity.
...
  GetContactos:
    type: object
    properties:
      id:
        type: integer
        format: int64
        example: 1
      nombre:
        type: string
        example: Sol Martinez
      email:
        type: string
        example: solmartinezpisani@email.com
      cargo: 
        type: string
        example: Dev Full-Stack
      interes:
        type: integer
        format: int64
        example: 25
      pais:
        type: string
        example: Argentina
      region:
        type: string
        example: Sudamérica
      compania:
        type: string
        example: Acamica
  GetContactoChannels: 
    type: object
    properties:
      canal:
        type: string
        example: Instagram
      canalId:
        type: integer
        format: int64
        example: 1
      usuario: 
        type: string
        example: solmartinez
      preferencia:
        type: string
        example: Sin Preferencia
  GetChannels: 
    type: object
    properties:
      id:
        type: integer
        format: int64
        example: 1
      nombre: 
        type: string
        example: Whatsapp
...
  CreateContacto:
    type: object
    properties:
      nombre:
        type: string
        example: Sol
      apellido:
        type: string
        example: Martinez
      email:
        type: string
        example: solmartinezpisani@email.com
      cargo: 
        type: string
        example: Dev Full-Stack
      interes:
        type: integer
        format: int64
        example: 25
      direccion: 
        type: string
        example: Av. Santa Fe 1444
      ciudad_id:
        type: integer
        format: int64
        example: 8
      compania_id:
        type: integer
        format: int64
        example: 5
...